Create Proteins Database
For this assignment, write a program that accesses and analyses protein sequences. The protein
sequences that you will analyse come from GenBank®, which is the National Institutes of Health
(USA) genetic sequence database1
. A full list of DNA, RNA, and protein sequences can be found
on their FTP site2
. For this assignment, use the abridged (protein_a.fa) and complete
(protein_c.fa) files provided on blackboard.

1. Selection of database. Create a user interface that helps navigate through the different
proteins. At the start of the program, the user should be asked whether he or she would like to use
the abridged (option 1) or complete (option 2) database. The selected database should then be
loaded into the program. (NB: Lecture 5 will describe key concepts that will be useful here)
2. Main menu. The user should then be presented with a menu of options.
1. Overview of this database. List the total number of proteins in the database and display a table
that associates each amino acid character to its proper name (e.g., ‘A: Alanine’).
2. Search by item id. This id refers to the protein in the order that it appears in the database.
3. Search by gi id. This id is the first identification labelled in the description line.
4. Search by ref id. This id is the second identification labelled in the description line.
5. Search by keyword. Search through the series of words within the third identification label of all
the proteins within the database for a keyword match. A match is made if the keyword is part of
another word or alone. If a match is found, display the protein details to the users and ask whether
this was the protein he or she was seeking. If multiple matches are found, display the options and
allow the user to choose which one to select or none of the above.
6. Quit the program.
3. Search hit menu. When a match is made, another set of options should be displayed.
1. Description. Display the gi id, the ref id, and the name of the protein.
2. Protein sequence. Display the full sequence.
3. Protein statistics. Display the number of amino acids in the sequence for each amino acid.
4. Record protein to file. Record the selected protein to a file called ‘selected_proteins.txt’.
5. Return to main menu.