Create Connect4
Specification
Connect Four is a game for two players, O and X, who take turns dropping discs from the top of a
seven-column, six-row vertically suspended grid. The pieces fall straight down, occupying the next
available space within the column. The objective of the game is to connect four of one's own discs
next to each other vertically, horizontally, or diagonally before your opponent. In this assignment,
you are tasked with coding this game.
Your program will prompt a specific player (O or X) to enter the column that they want to drop their
disc in. The program will evaluate whether the move is valid. If it is valid, the lowest available slot
of the column will be updated with a mark corresponding to the player’s symbol. If it is invalid, the
player will be asked to enter the column again. The program should run until Player X wins, Player
O wins, or when there’s a tie. A tie occurs when the board is full and neither player has won. The
program ends with the announcement of the game’s outcome.